Move GitHub modules under src/github
